requirements.txt
anyio==3.6.1
APScheduler==3.6.3
asgiref==3.5.2
asttokens==2.4.1
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
blinker==1.4
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloud-init==22.2
cloudinary==1.31.0
colorama==0.4.3
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.0
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.6
django-cors-headers==3.13.0
dnspython==2.2.1
entrypoints==0.3
executing==2.0.1
fiona==1.9.5
fonttools==4.44.1
geopandas==0.13.2
greenlet==3.0.3
h11==0.12.0
httpcore==0.15.0
httplib2==0.14.0
httpx==0.23.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.1
incremental==16.10.1
ipykernel==6.26.0
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.6.0
jupyter-core==5.5.0
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.7.3
matplotlib-inline==0.1.6
more-itertools==4.2.0
mysqlclient==2.1.1
nest-asyncio==1.5.8
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.1.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.18.0
pycryptodomex==3.19.0
Pygments==2.16.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.3.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyproj==3.5.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
pyzmq==25.1.1
requests==2.31.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
shapely==2.0.2
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.6.3
statsmodels==0.14.0
systemd-python==234
tornado==6.1
traitlets==5.13.0
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.14
wadllib==1.3.3
wcwidth==0.1.8
zipp==3.17.0
zope.interface==4.7.1
